+++ src/App.css
@@ -138,3 +138,14 @@ h2 {
   padding-top: 16px;
   border-top: 1px solid lightgray;
 }
+
+.no-todos-container {
+  width: 300px;
+  margin: 30px auto;
+  text-align: center;
+}
+
+.no-todos-container p {
+  margin-top: 20px;
+  text-align: center;
+}
