+++ src/components/App.jsx
@@ -1,6 +1,9 @@
 import '../reset.css'
 import '../App.css'
 import { useState } from 'react'
+import { NoTodos } from './NoTodos'
+import { TodoForm } from './TodoForm'
+import { TodoList } from './TodoList'
 
 function App() {
   const [todos, setTodos] = useState([
@@ -24,29 +27,20 @@ function App() {
     },
   ])
 
-  const [todoInput, setTodoInput] = useState('')
   const [idForTodo, setIdForTodo] = useState(4)
 
   // Function that adds new tasks
-  const addTodo = e => {
-    e.preventDefault()
-
-    // Don´t continue if input is empty string
-    if (todoInput.trim().length === 0) {
-      return
-    }
-
+  const addTodo = todo => {
     // Copy array and add a new task
     setTodos([
       ...todos,
       {
         id: idForTodo,
-        title: todoInput,
+        title: todo,
         isComplete: false,
       },
     ])
 
-    setTodoInput('')
     // Copy previouse idForTodo and increment + 1
     setIdForTodo(prevIdForTodo => prevIdForTodo + 1)
   }
@@ -116,99 +110,23 @@ function App() {
     setTodos(updatedTodos)
   }
 
-  // Function that handles icomming inputs
-  const handleInput = e => {
-    setTodoInput(e.target.value)
-  }
-
   return (
     <div className="todo-app-container">
       <div className="todo-app">
         <h2>Todo App</h2>
-        <form action="#" onSubmit={addTodo}>
-          <input
-            type="text"
-            value={todoInput}
-            onChange={e => handleInput(e)}
-            className="todo-input"
-            placeholder="What do you need to do?"
+        <TodoForm addTodo={addTodo} />
+        {todos.length > 0 ? (
+          <TodoList
+            todos={todos}
+            completeTodo={completeTodo}
+            markAsEditing={markAsEditing}
+            updateTodo={updateTodo}
+            cancelEdit={cancelEdit}
+            deleteTodo={deleteTodo}
           />
-        </form>
-
-        <ul className="todo-list">
-          {todos.map((todo, index) => (
-            <li key={todo.id} className="todo-item-container">
-              <div className="todo-item">
-                <input
-                  type="checkbox"
-                  onChange={() => completeTodo(todo.id)}
-                  checked={todo.isComplete ? true : false}
-                />
-                {!todo.isEditing ? (
-                  <span
-                    onDoubleClick={() => markAsEditing(todo.id)}
-                    className={`todo-item-label ${
-                      todo.isComplete ? 'line-through' : ''
-                    }`}
-                  >
-                    {todo.title}
-                  </span>
-                ) : (
-                  <input
-                    type="text"
-                    onBlur={e => updateTodo(e, todo.id)}
-                    onKeyDown={e => {
-                      if (e.key === 'Enter') {
-                        updateTodo(e, todo.id)
-                      } else if (e.key === 'Escape') {
-                        cancelEdit(e, todo.id)
-                      }
-                    }}
-                    className="todo-item-input"
-                    defaultValue={todo.title}
-                    autoFocus
-                  />
-                )}
-              </div>
-              <button onClick={() => deleteTodo(todo.id)} className="x-button">
-                <svg
-                  className="x-button-icon"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M6 18L18 6M6 6l12 12"
-                  />
-                </svg>
-              </button>
-            </li>
-          ))}
-        </ul>
-
-        <div className="check-all-container">
-          <div>
-            <div className="button">Check All</div>
-          </div>
-
-          <span>3 items remaining</span>
-        </div>
-
-        <div className="other-buttons-container">
-          <div>
-            <button className="button filter-button filter-button-active">
-              All
-            </button>
-            <button className="button filter-button">Active</button>
-            <button className="button filter-button">Completed</button>
-          </div>
-          <div>
-            <button className="button">Clear completed</button>
-          </div>
-        </div>
+        ) : (
+          <NoTodos />
+        )}
       </div>
     </div>
   )
