+++ src/components/NoTodos.jsx
@@ -1,0 +1,8 @@
+export const NoTodos = () => {
+  return (
+    <div className="no-todos-container">
+      <img src="/undraw_add_tasks_re_s5yj.svg" alt="Add Task" width={200} />
+      <p>Add some todos...</p>
+    </div>
+  )
+}
